api/requirements.txt
anyio==3.6.2
black==22.12.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
fastapi==0.88.0
h11==0.14.0
httptools==0.5.0
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.0
pydantic==1.10.2
Pygments==2.13.0
python-decouple==3.6
python-dotenv==0.21.0
PyYAML==6.0
rich==12.6.0
shellingham==1.5.0
sniffio==1.3.0
starlette==0.22.0
typer==0.7.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
